+++ .gitsplit.yml
@@ -25,11 +25,9 @@ splits:
   - prefix: "src/Contrib/Newrelic"
     target: "https://${GH_TOKEN}@github.com/opentelemetry-php/exporter-newrelic.git"
   - prefix: "src/Contrib/Otlp"
-    target: "https://${GH_TOKEN}@github.com/opentelemetry-php/exporter-otlp-common.git"
+    target: "https://${GH_TOKEN}@github.com/opentelemetry-php/exporter-otlp.git"
   - prefix: "src/Contrib/Grpc"
     target: "https://${GH_TOKEN}@github.com/opentelemetry-php/exporter-otlp-grpc.git"
-  - prefix: "src/Contrib/OtlpHttp"
-    target: "https://${GH_TOKEN}@github.com/opentelemetry-php/exporter-otlp-http.git"
   - prefix: "src/Contrib/Zipkin"
     target: "https://${GH_TOKEN}@github.com/opentelemetry-php/exporter-zipkin.git"
   - prefix: "src/Contrib/ZipkinToNewrelic"
