+++ src/Contrib/Otlp/composer.json
@@ -1,7 +1,7 @@
 {
-  "name": "open-telemetry/exporter-otlp-common",
-  "description": "OTLP common code for OpenTelemetry exporters.",
-  "keywords": ["opentelemetry", "otel", "tracing", "metrics", "exporter", "otlp", "grpc"],
+  "name": "open-telemetry/exporter-otlp",
+  "description": "OTLP exporter for OpenTelemetry.",
+  "keywords": ["opentelemetry", "otel", "tracing", "metrics", "exporter", "otlp", "grpc", "http"],
   "type": "library",
   "license": "Apache-2.0",
   "authors": [
@@ -12,9 +12,15 @@
   ],
   "require": {
     "php": "^7.4 || ^8.0",
+    "php-http/discovery": "^1.14",
     "open-telemetry/gen-otlp-protobuf": "self.version",
     "open-telemetry/api": "self.version",
     "open-telemetry/sdk": "self.version"
+  },
+  "suggest": {
+    "php-http/async-client-implementation": "For otlp/http exporting",
+    "php-http/discovery": "For otlp/http exporting",
+    "psr/http-factory-implementation": "For otlp/http exporting"
   },
   "autoload": {
     "psr-4": {
